pyproject.toml
[project]
name = "cte-project"
version = "0.1.0"
description = "Characteristics Traits Evaluator (CTE) – personal ML + LLM pipeline"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas (>=2.3.2,<3.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "scipy (>=1.16.1,<2.0.0)",
    "scikit-learn (>=1.3,<2)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "rich (>=14.1.0,<15.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "streamlit (>=1.48.1,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "typer (>=0.16.1,<0.17.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "transformers (>=4.55.3,<5.0.0)",
    "sentence-transformers (>=5.1.0,<6.0.0)",
    "vadersentiment (>=3.3.2,<4.0.0)",
    "xgboost (>=3.0.4,<4.0.0)",
    "ipykernel (>=6.30.1,<7.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "pyarrow (>=21.0.0,<22.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "statsmodels (>=0.14.5,<0.15.0)",
    "wordcloud (>=1.9.4,<2.0.0)",
    "shap (>=0.49.1,<0.50.0)",
    "textblob (>=0.19.0,<0.20.0)",
    "openai (>=2.7.1,<3.0.0)",
    "pillow (>=12.0.0,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.4.6"
ipykernel = "^6.30.1"
matplotlib = "^3.10.5"
seaborn = "^0.13.2"
statsmodels = "^0.14.5"
scipy = "^1.16.1"

[tool.poetry]
packages = [
  { include = "cte", from = "src" }
]
